v2

+++ DigitalCatalog/src/app/dashboard/dashboard.component.ts
@@ -2,6 +2,8 @@ import { Component, OnInit } from "@angular/core";
 import { AuthService } from "../services/auth.service";
 import User from "../models/user.model";
 import { Router } from "@angular/router";
+import { MatDialog } from "@angular/material/dialog";
+import { DialogBoxComponent } from "../dialog-box/dialog-box.component";
 
 @Component({
   selector: "app-dashboard",
@@ -11,7 +13,7 @@ import { Router } from "@angular/router";
 export class DashboardComponent implements OnInit {
   userData: User;
 
-  constructor(private authService: AuthService, private router: Router) {}
+  constructor(private authService: AuthService, private router: Router, private dialog: MatDialog) {}
 
   ngOnInit(): void {
     this.authService.userData.subscribe((res: User) => {
@@ -20,7 +22,15 @@ export class DashboardComponent implements OnInit {
   }
 
   onLogout(): void {
-    localStorage.clear();
-    this.router.navigate(["login"]);
+    const dialogRef = this.dialog.open(DialogBoxComponent, {
+      data: { name: `${this.userData.firstName}, are you sure you want to logout?` },
+    });
+
+    dialogRef.afterClosed().subscribe((res) => {
+      if (res === true) {
+        localStorage.clear();
+        this.router.navigate(["login"]);
+      }
+    });
   }
 }
